Redirect away from authentication pages if user is authenticated
Ran into an issue where even after entering the correct OTP, the user was shown the authentication page again. This may be because we are not using cookies to remember users (i.e., remember_otp_session_for_seconds is the default of 0). This solution is to perform a redirect to the `after_two_factor_success_path_for` if the user `is_fully_authenticated?`.